requirements.txt
trimesh==4.11.2
ezdxf==1.4.3
shapely==2.1.2
svg.path==7.0
gmsh==4.15.1
Pillow==12.1.1
numpy==2.4.2
